refactor(tui): extract SharedLayoutRuntime into tui::runtime
The 1.5k-line runtime leaves the module root and its impl is split by
concern: the node API, the terminal loop, layout intents, and input
forwarding, with the integration tests alongside the struct.
